Implement risk score checker using JSONs and protos.
I'm unsure what the original CSV files looked like, so impossible to say if I'm processing this data in a sane way yet. However, this computes a risk score for each participant to participant interaction in each experiment.

Need to next learn if this is being computed the right way and generate something that I can compare it to.

This also includes results.md which contains the output of the script for each viewing.